Assigning Bugs Fixed - Editor/Owner only perms added
you can now assign bugs to people/need to make it so its only people inside your org and if they aren't in your org to send a email out. Also added editor and only permissions to the assigning bugs and editing bug information so you cannot access it unless you are a org owner or editor.