V0.01 - Create Org / Org Display WIP
You can now create orgs and the org display is starting to get worked on now and will start displaying orgs that you can click on and view the overview of the org and projects. From the Org Display page you will be able to see info such as the member inside of the org if you have that permission level.