V0.01 - Org Display WIP Description View
Description is now showing and is responsive on mobile need to add the projects to display should be a easy thing to implement within 2 hours The project view will show the bugs that have been created on the project and then you can assign people to bugs from the page as well as you can read upon what the project is for and what it is used for. Im also planning on adding a private or public feature for your project depending wether its open source or not.